+++ packages/theme_tailor/lib/src/generator/tailor_annotations_generator.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: deprecated_member_use
 
-import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/element2.dart';
 import 'package:theme_tailor/src/generator/generator_annotation_matcher.dart';
 import 'package:theme_tailor/src/generator/tailor_mixin_generator.dart';
 import 'package:theme_tailor_annotation/theme_tailor_annotation.dart';
@@ -11,5 +11,5 @@ class TailorAnnotationsGenerator extends GeneratorAnnotationMatcher<TailorMixin>
   final TailorMixin buildYamlConfig;
 
   @override
-  GeneratorToBuffer<TailorMixin> getGeneratorFrom(Element element) => TailorMixinGenerator(buildYamlConfig);
+  GeneratorToBuffer<TailorMixin> getGeneratorFrom(Element2 element) => TailorMixinGenerator(buildYamlConfig);
 }