+++ packages/theme_tailor/lib/src/util/extension/element_extension.dart
@@ -1,4 +1,4 @@
-import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/element2.dart';
 import 'package:collection/collection.dart';
 import 'package:json_annotation/json_annotation.dart' show JsonSerializable;
 import 'package:source_gen/source_gen.dart';
@@ -6,14 +6,14 @@ import 'package:theme_tailor/src/model/constructor_data.dart';
 import 'package:theme_tailor/src/util/extension/parameter_element_extension.dart';
 import 'package:theme_tailor_annotation/theme_tailor_annotation.dart';
 
-extension ElementExtension on Element {
+extension ElementExtension on Element2 {
   bool get hasJsonSerializableAnnotation {
     const checker = TypeChecker.fromRuntime(JsonSerializable);
     return checker.hasAnnotationOf(this, throwOnUnresolved: false);
   }
 
   bool isFromPackage(String package) {
-    return library?.librarySource.fullName.startsWith('/$package/') ?? false;
+    return library2?.firstFragment.source.fullName.startsWith('/$package/') ?? false;
   }
 
   bool get hasTailorMixinAnnotation {
@@ -27,25 +27,25 @@ extension ElementExtension on Element {
   }
 }
 
-extension ClassElementExtensions on ClassElement {
+extension ClassElementExtensions on ClassElement2 {
   bool hasMixinNamed(String mixin) {
-    return mixins.map((e) => e.element.name).contains(mixin);
+    return mixins.map((e) => e.element3.name3).contains(mixin);
   }
 
-  ConstructorElement? get _defaultCtor {
-    return constructors.firstWhereOrNull((ctor) => ctor.name.isEmpty);
+  ConstructorElement2? get _defaultCtor {
+    return constructors2.firstWhereOrNull((ctor) => ctor.name3 == 'new');
   }
 
   ConstructorData? constructorData() {
     final ctor = _defaultCtor;
-    final parameters = ctor?.parameters;
+    final parameters = ctor?.formalParameters;
 
     if (parameters == null || parameters.isEmpty) return null;
 
     return ConstructorData(
       constructorName: ctor!.displayName,
       parameterNameToType: Map.fromEntries(
-        parameters.map((e) => MapEntry(e.name, e.parameterType)),
+        parameters.map((e) => MapEntry(e.name3!, e.parameterType)),
       ),
     );
   }