+++ packages/theme_tailor/lib/src/util/extension/parameter_element_extension.dart
@@ -1,7 +1,7 @@
-import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/element2.dart';
 import 'package:theme_tailor/src/model/constructor_parameter_type.dart';
 
-extension ParameterElementExtension on ParameterElement {
+extension ParameterElementExtension on FormalParameterElement {
   CtorParamType get parameterType {
     return isNamed
         ? CtorParamType.named